pyproject.toml
[tool.poetry]
name = "melody"
version = "0.0.1"
description = "Infra Orchestrator"
authors = ["Rohit Rai <rohit@aitomatic.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
pulumi = "^3"
pulumi-aws = "^4"
prefect = {extras = ["viz"], version = "^0.15.6"}
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
